feat: Add gauge card component
Add a new component called GaugeCard that displays a circular gauge with a percentage value and a label. The component is used to visualize data in a visually appealing way. The gauge card has a hover effect that lifts it up and adds a shadow. The gauge card is styled with a dark background color and white text. The gauge itself is a SVG element that shows the percentage value as a filled arc. The label is displayed below the gauge. This component will be used in various parts of the application to display data in a visually appealing manner.